* Tweaked victory menu * Damage is now calculated as float instead of int

+++ Assets/Scripts/Enemy/EnemyHealth.cs
@@ -37,9 +37,9 @@ public class EnemyHealth : MonoBehaviour
         return dead;
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
-        currentHealth -= amount;
+        currentHealth -= (int)amount;
 
         if (currentHealth <= 0 && !dead)
         {
@@ -47,7 +47,7 @@ public class EnemyHealth : MonoBehaviour
         }
     }
 
-    public void ApplyDot(float duration, float tickInterval, int tickDamage)
+    public void ApplyDot(float duration, float tickInterval, float tickDamage)
     {
         dotEffects.AddLast(new DotEffect(duration, tickInterval, tickDamage));
     }
